white-city-arbox/backend/requirements.txt
httpx>=0.27.0
supabase>=2.4.0
python-dateutil>=2.9.0
python-dotenv>=1.0.0
modal>=0.62.0
